Add a reference to google/pprof#351
We do have a parser for the atos output (in JS though, not Go), but that's really just based on my guesses as to what we need. https://github.com/google/pprof/issues/351 is the matching ticket on the pprof repository that asks for atos support when `gaddr2line` (binutils) or `llvm-symbolizer` (llvm) are not available.

We should watch that ticket, and consider whatever is done there to be "canonical": Our implementation must match that.

Note: This doesn't mean we cannot contribute our implementation -- but we shouldn't own the parsing knowledge here.